Create an own temporary storage pool for the ISO image
When specifying the path of the ISO image directly to virt-install with the
--disk option, virt-install will try to create a storage pool of its own
which may sometimes (eg. when virt-install faced an error) already be
existant and cause virt-install to fail. Thus we now create a storage pool
of our own and delete (= undefine) it again at program termination.

Note that it's perfectly ok for libvirt to have multiple storage pools
referring the same directory, so even if there is already a storage pool
referencing the directory of the ISO image there's no problem with that.