Implement own LookupVMByName() to work around buggy libvirt.lookupByName()
The python libvirt binding's lookupByName() function appears to be buggy
in that it causes error messages to be emitted to stderr if a certain
domain (VM) does not exist. We work around this by implementing an own
LookupVMByName() method that looks in the output of listAllDomains().

This also introduces the necessity for VMState() to return a new return
code when a domain (VM) is not defined at all.