Introduce video capturing of VM console output
This introduces a new command line option --video which, if specified,
enables the creation of a video file containing screenshots of the VM
console output, thus allowing ex-post inspection of the installation
process and its possible failure. --video is optional, but if specified
its argument must be the filename of the video to create (currently an
.avi file).

The installation monitoring implemented by virtinsttest and its plugins
is usually fast enough. Capturing VM console output in addition can make
this process too slow but this ultimately depends on the concrete machine
in question. You can influence this by choosing a suitable --video-framerate
option. The default is 5 fps, thus virtinsttest will take a screenshot every
1/5 second and the video will have a framerate of 5 fps.

If --video is enabled, the active OS plugin can emit statistics at loglevel
DEBUG regarding its current performance, eg. the number of log lines
processed per second. If this number is too low and virtinsttest erroneously
thinks that the installation got stuck, you might have specified a too high
--video-framerate.

The VM console can switch to different resolutions, eg. when the OS's
graphical installer is loaded, but videos must have a fixed resolution.
Thus --video-width and --video-height can be used to specify the
maximum expected console width and height (default: 1024x768).

The screenshots get a footer text added at their bottom, indicating a
timestamp, the current installation state and its timeout as well as
the log message from the last state transition. The font for the footer
can be changed using the --video-footerfont command line option, which
takes the path to a TrueType font as argument.

In addition, plugins can augment the video with additional footer data,
eg. the log lines currently processed, as implemented in the
OpenSUSE13xPlugin.