Snapshot the VM's disks if installation stage 1 was completed successfully
This adds
1. support for creating VM snapshots (of external kind, to be precise)
2. creating a snapshot if installation 1 was completed successfully
   (INSTSTATE_STAGE1_FINISHED was reached).

Note that libvirt's snapshot feature is a bit ugly. Internal snapshots
(storing the snapshots inside the same file as the disk being snapshotted)
would be fully supported but are slow, external snapshots can only be
created, not deleted, for whatever reasons. In our case that's not too
bad, we just delete the disk files at teardown so we exit cleanly.